Group FUSE stuff into own submodule
